// Web/src/generated/model/definitionLifecycle.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ReleaseHub API
 * Shared HTTP API contract for the ReleaseHub frontend and backend.
 * OpenAPI spec version: 1.0.0
 */

export type DefinitionLifecycle = typeof DefinitionLifecycle[keyof typeof DefinitionLifecycle];


export const DefinitionLifecycle = {
  Draft: 'Draft',
  Published: 'Published',
  Disabled: 'Disabled',
} as const;